Simplify Go codebase: remove dead code, fix types, add test coverage
Recipe: Comb through the entire Go codebase (thermal/) with the spirit
of /simplify. Identify dead code, unused exports, redundant type casts,
duplicated functions, missing test coverage, and inconsistent patterns.
Fix everything that's low-to-medium effort. Specifically: remove unused
methods on collector types (TotalProcs, TotalRSS, TotalCPU, SwapPercent),
unused UI constants (SessionGlyph, CategoryGlyphDefault, AgentGlyphWidth,
threshold consts), unused formatFixedCount, unused StaleDataThreshold,
dead --horizontal/--vertical CLI flags, and the non-buffered sparkline
rendering path. Collapse smoothMap/smoothMapPerKey into a single smoothEMA
with an alpha function parameter. Make HeadroomCritBytes/HeadroomWarnBytes
and SwapWarmBytes/SwapHotBytes/SwapCritBytes proper int64 byte values
instead of requiring callers to multiply by GB. Remove redundant int64()
casts on untyped constants. Add a numCategories drift guard in rates init.
Rewrite brailletext to import RenderBrailleWord from widgets instead of
duplicating the font. Add HandleEvent tests for all event types, layout
compositor tests, and a slow-loop concurrency test for the collector.

Changes:
  thermal/internal/collector/types.go — Removed four unused methods (TotalProcs, TotalRSS, TotalCPU, SwapPercent) that had zero callers across the codebase.
  thermal/internal/ui/colors.go — Removed nine unused constants (SessionGlyph, CategoryGlyphDefault, AgentGlyphWidth, SpawnWarn/Crit, NetWarn/Crit, TotalWarn/Crit) left over from earlier widget designs.
  thermal/internal/widgets/rates.go — Removed unused formatFixedCount, removed redundant int64 cast, added init-time panic guard for numCategories drift against collector.Categories.
  thermal/internal/widgets/rates_test.go — Removed test for deleted formatFixedCount.
  thermal/internal/config/tuning.go — Removed unused StaleDataThreshold. Typed swap thresholds as int64. Changed HeadroomCritBytes/WarnBytes to byte values (2<<30, 4<<30) so callers no longer multiply by GB.
  thermal/cmd/thermal/main.go — Removed dead --horizontal and --vertical flag definitions.
  thermal/internal/model/threat.go — Removed swap const aliases and redundant int64() casts; uses config.Swap*Bytes directly.
  thermal/internal/layout/horizontal.go — Removed redundant int64(model.GB) casts.
  thermal/internal/model/projection.go — Updated HeadroomCritBytes/WarnBytes call sites to use new byte-valued constants directly.
  thermal/internal/model/projection_test.go — Updated headroom test to match new byte-valued constants.
  thermal/internal/model/state.go — Collapsed smoothMap and smoothMapPerKey into single smoothEMA with alpha function parameter. Extracted countAlpha as named function to match catAlpha pattern. Updated headroom alert threshold to use byte-valued config.
  thermal/internal/model/state_test.go — Added HandleEvent tests for GateCap, AgentStop alert, ParallelEngaged, and ParallelDisengaged event types.
  thermal/internal/widgets/sparkline.go — Removed non-buffered rendering path (RenderSparkline, prepareSparkData, prepareSparkMask, interpolateData, interpolateMask). Made buf required in renderSparklineCore.
  thermal/internal/widgets/sparkline_test.go — Removed tests for deleted non-buffered functions and the matches-unbuffered comparison test.
  thermal/internal/widgets/braillefont.go — Added doc comments for font bit tables. Symbols remain private since only RenderBrailleWord needs to be exported.
  thermal/cmd/brailletext/main.go — Rewrote to import widgets.RenderBrailleWord instead of duplicating font data. Replaced string concatenation with strings.Builder.
  thermal/internal/collector/collector_test.go — Added TestRunWith_SlowLoopRunsConcurrently verifying network check and slow stats collection run in parallel.
  thermal/internal/layout/horizontal_test.go — New file with 8 tests covering zero-size, idle view, active view, linecount-matches-height, collapse toggle, help toggle, and notification bar visibility.

+++ thermal/internal/layout/horizontal.go
@@ -183,8 +183,8 @@ func (h *Horizontal) idleView() string {
 	// System stats while idle
 	if h.state.Current != nil && h.height >= 4 {
 		snap := h.state.Current
-		memGB := snap.System.MemUsedBytes / int64(model.GB)
-		totalGB := snap.System.MemTotalBytes / int64(model.GB)
+		memGB := snap.System.MemUsedBytes / model.GB
+		totalGB := snap.System.MemTotalBytes / model.GB
 		stats := ui.DimText(fmt.Sprintf(" CPU:%03d%%  MEM:%02d/%02dGB", int(snap.System.CPUPercent), memGB, totalGB))
 		lines = append(lines, stats)
 	}
